Modified to edit or delete only allowed by its creator

+++ itemcatalog.py
@@ -341,6 +341,11 @@ def editCategory(category_id):
         Category).filter_by(id=category_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+    if categoryToEdit.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+            are not authorized to edit this category.\
+            Please create your own category in order\
+            to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             categoryToEdit.name = request.form['name']
@@ -358,6 +363,11 @@ def deleteCategory(category_id):
         Category).filter_by(id=category_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+    if categoryToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+            are not authorized to delete this category.\
+            Please create your own category in order\
+            to delete.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(categoryToDelete)
 
@@ -423,6 +433,11 @@ def editItem(item_id):
         return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
     categories = session.query(Category).order_by(asc(Category.name)).all()
+    if item.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+         are not authorized to edit this item.\
+          Please create your own item in order\
+           to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             item.name = request.form['name']
@@ -456,6 +471,11 @@ def deleteItem(item_id):
     if 'username' not in login_session:
         return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if item.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You\
+            are not authorized to delete this item.\
+            Please create your own item in order\
+            to delete.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(item)
         session.commit()
